Remove requirement on contact information.
If you access the crates.io API, there is a scraping policy that asks
you to set contact information in the user agent. However,
Panamax bypasses the API and grabs crates directly from the
underlying CDN. This does not require any user agent changes as it is
directly pulling from (as of the time of this commit) Amazon S3.