Fix Docker Image Failing to Sync (#62)
* Add git to the docker image

This is needed for the docker image to sync and serve packages
correctly. Without it panamax will not download most of the crates and
will throw 500 errors.

* Added '--rm' flag to docker example

Since the image's state is mounted into the container setting the
container to auto delete when it's process exits helps cut down on the
number of stopped images.

* Update readme with working build instructions

See issue #61

https://github.com/panamax-rs/panamax/issues/61

The command `cargo install panamax` fails because it doesn't default to
using locked files.