Use cargo vendor for mirroring selected crates
Add give a path to a folder created from cargo-vendor, which contains a folder of
all crates that can be easily parsed while letting cargo-vendor do the
hard job of grabbing the correct depends.

The code then walks that directory looking for crate name and the version, and only downloads
those specific crates from crates.io.